Move UnreferencedPeerState to own file
